requirements.txt
pymongo==4.17.0
python-dotenv==1.2.2
uvicorn==0.46.0
streamlit==1.37.1
requests==2.32.3
dnspython==2.8.0
pydantic==2.13.3
fastapi==0.136.1